Bug: control-level solved.

+++ Front-End/CSL_App/client/cslmain/cslmain.js
@@ -297,9 +297,16 @@ function layerExists(pControlNumb){
     return false;
 }
 
+function sourceExists(pControlNumb){
+    if(GlobalAppState.map.getSource(CONTROL_LIST[pControlNumb][CONTROL_ID]) != undefined)
+        return true;
+    return false;
+}
+
 function deleteLayer(pControlNumb){
     GlobalAppState.map.removeLayer(CONTROL_LIST[controlLevelNumber][CONTROL_ID]);
-    GlobalAppState.map.removeSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID]);
+    if(sourceExists(pControlNumb))
+        GlobalAppState.map.removeSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID]);
 }
 
 function createLayer(pControlNumb,pData){
@@ -309,32 +316,33 @@ function createLayer(pControlNumb,pData){
 }
 
 function setDataOnLayer(pData){
-    console.log('pData:')
-    console.log(pData)
     //console.log(pData)
     if (layerExists(controlLevelNumber)){
         if (isValidatedLevel(pData,controlLevelNumber)){
             deleteLayer(controlLevelNumber);
             createLayer(controlLevelNumber,pData);
-            setLevelController();
+            return true;
         }
     }
     else{
         if (isValidatedLevel(pData,controlLevelNumber)){
             createLayer(controlLevelNumber,pData);
-            setLevelController();
+            return true;
         }
     }
+    return false;
 }
 
 //======================Parser GEOJSON========================================================================
-
 function parsingMapJSON() {
     jsonInfo={};
     jsonInfo.layers = GlobalAppState.project.project_instance.layers ? GlobalAppState.project.project_instance.layers : []
     layer = {}
     layer.level=controlLevelNumber
     currSource= GlobalAppState.map.getSource(CONTROL_LIST[layer.level][CONTROL_ID]);
+    console.log("lawea ya sirve");
+    console.log(GlobalAppState.map.getSource(CONTROL_LIST[layer.level][CONTROL_ID])._data);
+
     if(currSource!=undefined)
     {
         layer.stages=[]
@@ -358,7 +366,18 @@ function parsingMapJSON() {
         else
             jsonInfo.layers.push(layer)
     }
+
+    console.log("jsonInfo");
+    console.log(jsonInfo);
     return jsonInfo;
+}
+
+function deleteAllMap(){
+    for(counter=0;counter< CONTROL_LIST.length;counter++){
+        if (layerExists(counter)){
+            deleteLayer(controlLevelNumber);
+        }
+    }
 }
 
 export function loadProject(pProject){
@@ -372,12 +391,13 @@ export function loadProject(pProject){
             features:[]
         });
     }
-    
+    let currLevel=0;
     for(var counter=0;counter<pProject.project_instance.layers.length;counter++){
         for(var counter2=0;counter2<pProject.project_instance.layers[counter].stages.length;counter2++){
             var feature=new Object();
             feature.properties=new Object();
             feature.type="Feature";
+            currLevel=pProject.project_instance.layers[counter].level;
             feature.id="c"+counter+"cc"+counter2+"level"+pProject.project_instance.layers[counter].level;
             var geometry=new Object();
             geometry.type="Polygon";
@@ -392,6 +412,7 @@ export function loadProject(pProject){
             levelList[pProject.project_instance.layers[counter].level].features.push(feature);
         }
     }
+
     //Pintar los niveles
     for(counter=0;counter<CONTROL_LIST.length;counter++){
         controlLevelNumber=counter;
@@ -399,6 +420,9 @@ export function loadProject(pProject){
             setDataOnLayer(levelList[counter]);
         }
     }
+    controlLevelNumber=currLevel;
+    //updateUIControlLevel(controlLevelNumber);
+    //setControlDraw(controlLevelNumber);
 }
 
 //======================Sistema de control========================================================================
@@ -419,9 +443,9 @@ function showButtonOnMapLayer(pElement){
     }
 }
 
-function setControlDraw(pControlDrawId){
-    controlDrawingId = pControlDrawId;
-    showButtonOnMapLayer(pControlDrawId);
+function setControlDraw(pControlLevel){
+    control=CONTROL_LIST[pControlLevel][CONTROL_TYPE];
+    showButtonOnMapLayer(control);
 }
 
 //==============================Meteor========================================================================
@@ -452,7 +476,7 @@ Template.CSL.events({
     'click #cslplano'(event, instance) {
         event.preventDefault();
         controlLevelNumber = 0;
-        setControlDraw(0);
+        setControlDraw(controlLevelNumber);
         if(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])!=undefined)
         ids = GlobalAppState.draw.set(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])._data);
         updateUIControlLevel(controlLevelNumber);
@@ -460,7 +484,7 @@ Template.CSL.events({
     'click #cslbloqueo'(event, instance) {
         event.preventDefault();
         controlLevelNumber = 1;
-        setControlDraw(0);
+        setControlDraw(controlLevelNumber);
         if(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])!=undefined)
         ids = GlobalAppState.draw.set(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])._data);
         updateUIControlLevel(controlLevelNumber);
@@ -468,7 +492,7 @@ Template.CSL.events({
     'click #cslacomet'(event, instance) {
         event.preventDefault();
         controlLevelNumber = 2;
-        setControlDraw(1);
+        setControlDraw(controlLevelNumber);
         if(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])!=undefined)
         ids = GlobalAppState.draw.set(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])._data);
         updateUIControlLevel(controlLevelNumber);
@@ -476,7 +500,7 @@ Template.CSL.events({
     'click #cslmaq'(event, instance) {
         event.preventDefault();
         controlLevelNumber = 3;
-        setControlDraw(0);
+        setControlDraw(controlLevelNumber);
         if(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])!=undefined)
         ids = GlobalAppState.draw.set(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])._data);
         updateUIControlLevel(controlLevelNumber);
@@ -484,7 +508,7 @@ Template.CSL.events({
     'click #cslconstrucc'(event, instance) {
         event.preventDefault();
         controlLevelNumber = 4;
-        setControlDraw(0);
+        setControlDraw(controlLevelNumber);
         if(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])!=undefined)
         ids = GlobalAppState.draw.set(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])._data);
         updateUIControlLevel(controlLevelNumber);
@@ -492,7 +516,7 @@ Template.CSL.events({
     'click #csltemp'(event, instance) {
         event.preventDefault();
         controlLevelNumber = 5;
-        setControlDraw(0);
+        setControlDraw(controlLevelNumber);
         if(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])!=undefined)
         ids = GlobalAppState.draw.set(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])._data);
         updateUIControlLevel(controlLevelNumber);
@@ -500,7 +524,7 @@ Template.CSL.events({
     'click #cslcivil'(event, instance) {
         event.preventDefault();
         controlLevelNumber = 6;
-        setControlDraw(1);
+        setControlDraw(controlLevelNumber);
         if(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])!=undefined)
         ids = GlobalAppState.draw.set(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])._data);
         updateUIControlLevel(controlLevelNumber);
@@ -509,7 +533,7 @@ Template.CSL.events({
         event.preventDefault();
         controlLevelNumber = 7;
         controlInputRadius = 20; //esto tiene que ser una entrada.
-        setControlDraw(2);
+        setControlDraw(controlLevelNumber);
         if(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])!=undefined)
         ids = GlobalAppState.draw.set(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])._data);
         updateUIControlLevel(controlLevelNumber);
@@ -518,7 +542,7 @@ Template.CSL.events({
         event.preventDefault();
         //Aqui hay que establecer las diferencias entre los poligonos para que queden los sectores libres.
         controlLevelNumber = 8;
-        setControlDraw(0);
+        setControlDraw(controlLevelNumber);
         if(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])!=undefined)
         ids = GlobalAppState.draw.set(GlobalAppState.map.getSource(CONTROL_LIST[controlLevelNumber][CONTROL_ID])._data);
         updateUIControlLevel(controlLevelNumber);
@@ -526,33 +550,39 @@ Template.CSL.events({
     //Boton arriba derecha
     'click #cslsavestruct'(event, instance) {
         event.preventDefault();
-        setDataOnLayer(GlobalAppState.draw.getAll());//establece los datos en el layer
-        updateUIControlLevel(controlLevelNumber);//Muestra en ui nivel actual
-        GlobalAppState.draw.deleteAll();
-        GlobalAppState.project.project_instance.layers = parsingMapJSON().layers;
-        Meteor.call("saveProject", GlobalAppState.project,
-            (error, result) => {
-                //console.log(result)
-                if(result.code > 0){
-                    Meteor.call("getProject", {key:GlobalAppState.project.key},
-                        (error, result) => {
-                            console.log(result)
-                            if(result !== undefined){
-                                GlobalAppState.project.project_instance = result;                     
-                                notify("Proyecto guardado", 3000, 'rounded')
-                            }
-                            else{
-                                notify("Error guardando el proyecto", 3000, 'rounded')        
-                            }
+        if(CONTROL_LIST.length>controlLevelNumber){
+            isDataSet=setDataOnLayer(GlobalAppState.draw.getAll());//establece los datos en el layer
+            if(isDataSet){
+                GlobalAppState.project.project_instance.layers = parsingMapJSON().layers;
+                Meteor.call("saveProject", GlobalAppState.project,
+                    (error, result) => {
+                        //console.log(result)
+                        if(result.code > 0){
+                            Meteor.call("getProject", {key:GlobalAppState.project.key},
+                                (error, result) => {
+                                    console.log(result)
+                                    if(result !== undefined){
+                                        GlobalAppState.project.project_instance = result;                     
+                                        notify("Proyecto guardado", 3000, 'rounded')
+                                    }
+                                    else{
+                                        notify("Error guardando el proyecto", 3000, 'rounded')        
+                                    }
+                                }
+                            )
+                            //notify("Project: " + result.state.message, 3000, 'rounded')
                         }
-                    )
-                    //notify("Project: " + result.state.message, 3000, 'rounded')
-                }
-                else{
-                    //notify("Project: " + result.state.message, 3000, 'rounded')        
-                }
+                        else{
+                            //notify("Project: " + result.state.message, 3000, 'rounded')        
+                        }
+                    }
+                )
+                setLevelController();
+                updateUIControlLevel(controlLevelNumber);//Muestra en ui nivel actual
+                setControlDraw(controlLevelNumber);
+                GlobalAppState.draw.deleteAll();
             }
-        )
+        }
     }
 })
 
@@ -584,7 +614,7 @@ Template.CSL.onRendered(
         });
         
         GlobalAppState.map.addControl(GlobalAppState.draw);//add controls on mapbox, set the draw tool
-        setControlDraw(0);//Set the button of the first level
+        setControlDraw(controlLevelNumber);//Set the button of the first level
 
         /*console.log("Sistema de geolocalizador 1")
         MapboxGeocoder.query('Montreal Quebec');
@@ -607,17 +637,17 @@ Template.CSL.onRendered(
                 return layer.type !== 'symbol';
             });
             
-            GlobalAppState.map.on('dblclick', 'cslplano', function (e) {
+            /*GlobalAppState.map.on('dblclick', 'cslplano', function (e) {
                 polygon = turf.polygon(e.features[0].geometry.coordinates);
                 center = turf.centerOfMass(polygon);
                 var area = turf.area(polygon);
                 //console.log("CENTER:"+center.geometry.coordinates);
                 notify("Area: "+area+" m2", 3000, 'rounded')
-                /*new mapboxgl.Popup()
+                new mapboxgl.Popup()
                     .setLngLat(center.geometry.coordinates)
                     .setHTML("<h4>Here is a ne element</h4>")
-                    .addTo(map);*/
-            });
+                    .addTo(map);
+            });*/
             /*      map.on('mouseenter', 'cslplano', function () {
                     map.getCanvas().style.cursor = 'pointer';
                 });
